commands-file: added `bt of the oldest blocked process`:
this was a tricky one, because running bt on the blocked process
requires reducing the ps list to the desired PID and invoking
bt with that PID. To my knowledge, there is no way that we can store
variables in crash, and `bt` does not accept piped inputs either. so
we needed to hack our way around.

this implementation of `bt of the oldest proc` leverages one little posix
guarantee that programmers hate: concurrent reads and writes are fine on a
file and any read call must reflect the write calls prior. so, we can read
from a file and let crash execute them where the read commands are also
appending new commands to the file. this works just fine because crash reads
the input file progressively.

Signed-off-by: Mustafa Kemal Gilor <mustafa.gilor@canonical.com>