ci: migrate ubuntu:18.04 jobs to docker
unfortunately for us, github actions has deprecated the support for 18.04
runners and they seem to be no longer available. This patch migrates the 18.04
jobs to run on 18.04 docker container in order to keep test coverage for
bionic.

also removed "push" trigger from ci workflow (seemed a little bit excessive) and
set the default shell to "bash" for "install dependencies" job.

Signed-off-by: Mustafa Kemal Gilor <mustafa.gilor@canonical.com>